requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Async HTTP Client
httpx==0.25.1
aiofiles==23.2.1

# Document Processing
python-multipart==0.0.6
python-magic==0.4.27
PyPDF2==3.0.1
pillow==10.1.0

# Data Validation & Serialization
jsonschema==4.20.0

# Database & Caching
redis==5.0.1
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# Task Queue
celery[redis]==5.3.4

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx-mock==0.4.0

# Development
python-dotenv==1.0.0
black==23.11.0
ruff==0.1.6
mypy==1.7.1